Refactor: "AuthService 인터페이스" 분리
인터페이스로 분리
